+++ component/backend/sql/install/mysql/install.sql
@@ -3,6 +3,7 @@ CREATE TABLE IF NOT EXISTS `#__contactus_categories` (
   `title` varchar(255) NOT NULL DEFAULT '',
   `email` varchar(255) DEFAULT '',
   `sendautoreply` tinyint(3) DEFAULT 0,
+  `ticketcategory` BIGINT(20) UNSIGNED NOT NULL DEFAULT 0,
   `autoreply` text,
   `access` int(5) NOT NULL DEFAULT '1',
   `language` varchar(50) NOT NULL DEFAULT '*',