+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1
+
+* Version is 1.1.1 instead of 1.1.0 because of unfortunate SQL file naming back in 2013
+* Post as ATS ticket when possible (configurable per contact category)
+
 # 1.0.0
 
 First release with a version number.